slack_bot/requirements.txt
slack-bolt>=1.18.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
asyncio
typing-extensions>=4.0.0